Move tests to /tests
To stop my rust-anlayzer from killing me
